requirements.txt
PyQt6
sounddevice
webrtcvad-wheels
openai
pynput
numpy
mss
Pillow